Rewrite this code to use pint, a units framework for python
This allows us to prove that there are no units errors!
There were none, so the output is unchanged.

I used an extra set of quantities for Mpc/h and Msun/h, so I can also
check the reduced-h quantities needed for the halo mass function.

As well as a cross-check, this is a test case for using pint in
projects.